[Dashboard] Add contest and expedition cards
